Rename layouts and schemas for existing wizard
