+++ libs/newsletter-workflow/src/lib/executeCreate.ts
@@ -11,7 +11,7 @@ export const executeCreate: AsyncExecution = async (
 	stepLayout,
 	storageInstance,
 ): Promise<WizardFormData | string> => {
-	const schema = formSchemas['createNewsletter']; // TODO - this needs to be generalised
+	const schema = formSchemas['startDraftNewsletter']; // TODO - this needs to be generalised
 	if (!storageInstance) {
 		throw new Error('no storageInstance');
 	}
