+++ libs/newsletter-workflow/src/lib/newsletter-workflow.ts
@@ -1,16 +1,16 @@
 import type { WizardLayout } from '@newsletters-nx/state-machine';
 import { brazeLayout } from './steps/brazeLayout';
 import { cancelLayout } from './steps/cancelLayout';
-import { createNewsletterLayout } from './steps/createNewsletterLayout';
+import { createDraftNewsletterLayout } from './steps/createDraftNewsletterLayout';
 import { descriptionLayout } from './steps/descriptionLayout';
-import { editDraftNewsletterLayout } from './steps/editDraftNewsletter';
+import { editDraftNewsletterLayout } from './steps/editDraftNewsletterLayout';
 import { finishLayout } from './steps/finishLayout';
 import { identityNameLayout } from './steps/identityNameLayout';
 import { ophanLayout } from './steps/ophanLayout';
 import { pillarLayout } from './steps/pillarLayout';
 
 export const newslettersWorkflowStepLayout: WizardLayout = {
-	createNewsletter: createNewsletterLayout,
+	createDraftNewsletter: createDraftNewsletterLayout,
 	editDraftNewsletter: editDraftNewsletterLayout,
 	cancel: cancelLayout,
 	identityName: identityNameLayout,
